+++ apps/frontend-v3/app/(app)/portfolio/[chain]/page.tsx
@@ -15,7 +15,7 @@ import { Button, Card, HStack, Heading, Skeleton, Stack, Text, VStack } from '@c
 import { capitalize } from 'lodash'
 import { useParams } from 'next/navigation'
 import { useState } from 'react'
-import { getUserReferenceTokens } from '@repo/lib/modules/pool/pool.tokens.utils'
+import { getUserReferenceTokens } from '@repo/lib/modules/pool/pool-tokens.utils'
 import { PoolCore } from '@repo/lib/modules/pool/pool.types'
 
 export default function NetworkClaim() {
