+++ packages/lib/modules/pool/PoolList/PoolListTable/PoolListTableRow.tsx
@@ -13,7 +13,7 @@ import { getPoolPath } from '../../pool.utils'
 import { getUserTotalBalanceUsd } from '../../user-balance.helpers'
 import { usePoolList } from '../PoolListProvider'
 import { PoolListTokenPills } from '../PoolListTokenPills'
-import { getUserReferenceTokens } from '../../pool.tokens.utils'
+import { getUserReferenceTokens } from '../../pool-tokens.utils'
 
 interface Props extends GridProps {
   pool: PoolListItem
