+++ packages/lib/modules/pool/PoolProvider.tsx
@@ -17,7 +17,7 @@ import { usePoolEnrichWithOnChainData } from './queries/usePoolEnrichWithOnChain
 import { useOnchainUserPoolBalances } from './queries/useOnchainUserPoolBalances'
 import { useInvalidVariantRedirect } from './pool.hooks'
 import { useTokens } from '../tokens/TokensProvider'
-import { getCompositionTokens } from './pool.tokens.utils'
+import { getCompositionTokens } from './pool-tokens.utils'
 
 // TODO: replace all references to use Pool type in pool.types.ts in another PR
 export type Pool = GetPoolQuery['pool']
